rename pkg/api to pkg/instancedeleter
