+++ docs/instance-deleter.md
